:recycle: Refactor: 그룹 관련 API 위치 이동
<body>
- :external-api:/api/group

<footer>
- 관련: #145
