+++ application/external-api/src/main/java/com/namo/spring/application/external/api/group/dto/GroupResponse.java
@@ -1,0 +1,51 @@
+package com.namo.spring.application.external.api.group.dto;
+
+import java.util.List;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+public class GroupResponse {
+	private GroupResponse() {
+		throw new IllegalStateException("Util Class");
+	}
+
+	@AllArgsConstructor
+	@Getter
+	@Builder
+	public static class GroupIdDto {
+		private Long groupId;
+	}
+
+	@Getter
+	@AllArgsConstructor
+	@Builder
+	public static class GroupDto {
+		private Long groupId;
+		private String groupName;
+		private String groupImgUrl;
+		private String groupCode;
+		private List<GroupUserDto> groupUsers;
+	}
+
+	@Getter
+	@AllArgsConstructor
+	@NoArgsConstructor
+	@Builder
+	public static class GroupUserDto {
+		private Long userId;
+		private String userName;
+		private Integer color;
+	}
+
+	@Getter
+	@AllArgsConstructor
+	@NoArgsConstructor
+	@Builder
+	public static class GroupParticipantDto {
+		private Long groupId;
+		private String code;
+	}
+}
